refactor(example): reset the path of module import
