+++ example/single.ts
@@ -1,4 +1,4 @@
-import { TonHandler } from '../packages/ton/src'
+import { TonHandler } from '../packages/ton/dist'
 
 const single: TonHandler = () => ''
 export default single
